request for comments on contributing standards and acceptance criteria
As the pull requests pile up, and as contributors may have motives that stray from the purpose of helping other developers (promotion, spam, etc.), I wanted to start a conversation on what the criteria should be for accepting new contributions. To start it off, here are some things I thought of:

* What falls under the scope of developer tools? If it's outside (e.g. DApps) should we include a separate heading for well-built DApps that aid people's understanding of them
* Should proposed tools have met a certain threshold of usage? Could this be captured in a number of Github stars or forks?
* Should repo maintainers be required to use and evaluate the tools before adding/removing from the list?
* What reasons/feedback should be provided by maintainers if a proposed tool is rejected?

These are the opinions of myself, not ConsenSys.